+++ src/sgpp/pde/application/PoissonEquationSolver.hpp
@@ -56,7 +56,7 @@ public:
 	 */
 	virtual ~PoissonEquationSolver();
 
-	void constructGrid(sg::base::BoundingBox& myBoundingBox, size_t level);
+	void constructGrid(sg::base::BoundingBox& myBoundingBox, int level);
 
 	void solvePDE(sg::base::DataVector& alpha, sg::base::DataVector& rhs, size_t maxCGIterations, double epsilonCG, bool verbose = false);
